+++ src/lib/alphaIntegration/index.ts
@@ -1,0 +1,37 @@
+export {
+  SPRINT103_ALPHA_INTEGRATION_VERSION,
+  type AlphaJourneyStageId,
+  type AlphaJourneyRouteId,
+  type AlphaJourneyRouteDef,
+  type AlphaJourneyStageDef,
+  type AlphaIntegrationFlagReport,
+  type AlphaIntegrationDegradation,
+} from './types'
+export {
+  ALPHA_JOURNEY_ROUTES,
+  resolveJourneyPath,
+  listKnownJourneyPaths,
+} from './routes'
+export {
+  ALPHA_JOURNEY_STAGES,
+  allStagesConnected,
+} from './stages'
+export {
+  bookingComposeFromAgentMeta,
+  resolveAlphaNextStep,
+  type AlphaNextStep,
+} from './dataFlow'
+export {
+  ALPHA_INTEGRATION_FLAG_IDS,
+  reportAlphaIntegrationFlags,
+  isAlphaIntegrationFlagEnabled,
+} from './flags'
+export {
+  degradationForMissing,
+  safeDegradationMessage,
+} from './degradation'
+export {
+  buildAlphaIntegrationReport,
+  resolveBookingEntryPath,
+  type AlphaIntegrationReport,
+} from './composer'